WIP on intermediate representation
The existing layout algorithm does not take account of line width when
positioning images. This change adds an intermediate representation that
records for every path it's drawing context thereby allowing us to
include line width in layout calculations.

This change will also fix the handling of `at` layout.
