+++ shared/src/main/scala/doodle/core/Image.scala
@@ -30,14 +30,6 @@ sealed trait Image {
 
   def fillColor(color: Color): Image =
     ContextTransform(_.fillColor(color), this)
-
-  /** Get a bounding box around this Image.
-    *
-    * Implemented here so we can cache the results, and avoid frequently
-    * recomputing the bounding box
-    */
-  lazy val boundingBox: BoundingBox =
-    BoundingBox(this)
 }
 object Image {
